+++ test/bdd-syntax/shared/parameterTypes.ts
@@ -1,9 +1,14 @@
 import { defineParameterType } from 'playwright-bdd';
 
-type Color = 'red' | 'blue' | 'yellow';
+export type Color = 'red' | 'blue' | 'yellow';
 
 defineParameterType({
   name: 'color',
   regexp: /red|blue|yellow/,
   transformer: (s) => s.toLowerCase() as Color,
+});
+
+defineParameterType({
+  name: 'param-no-transformer',
+  regexp: /"(dog|cat)"/,
 });